+++ template_game/src/main.rs
@@ -1,21 +1,38 @@
-//! Everything needed to run the main game logic 
+//! Everything needed to run the main game logic
 
-use bevy::prelude::*;
+use bevy::{
+    prelude::*,
+    render::{
+        settings::{Backends, WgpuSettings},
+        RenderPlugin,
+    },
+};
 
-/// This lets us set a state to select from for system scheduling.  
-#[derive(Clone, Debug, Default, Eq, PartialEq, Hash, States)]
+/// Set the game state to align systems with their respective runtimes
+#[derive(Debug, Clone, Copy, Default, Eq, PartialEq, Hash, States)]
 enum GameState {
-    // TODO: Choose desired default
-    MainMenu,
     #[default]
+    Menu,
     Playing,
-    PauseMenu,
-
+    GameOver,
 }
 
 fn main() {
-    App::new()
-        .add_plugins(DefaultPlugins)
-        .add_state::<GameState>()
-        .run();
+    App::new().add_plugins(
+        DefaultPlugins
+            .set(WindowPlugin {
+                primary_window: Some(Window {
+                    title: String::from("3D Bevy Template"),
+                    ..default()
+                }),
+                ..default()
+            })
+            // Work around for https://github.com/bevyengine/bevy/issues/7620
+            .set(RenderPlugin {
+                wgpu_settings: WgpuSettings {
+                    backends:Some(Backends::PRIMARY),
+                    ..Default::default()
+                },
+            }),
+        ).run();
 }