+++ template_lib/src/lib.rs
@@ -1,0 +1,9 @@
+//! Provides plugins needed by the TEMPLATE game.
+#![deny(missing_docs)]
+#![deny(clippy::missing_docs_in_private_items)]
+#![forbid(unsafe_code)]
+#![warn(clippy::doc_markdown)]
+// Often exceeded by queries
+#![allow(clippy::type_complexity)]
+
+pub mod player;