+++ template_lib/src/player/camera.rs
@@ -1,0 +1,12 @@
+//! Code needed to run the game camera
+
+use bevy::prelude::{App, Plugin, Commands};
+
+/// Camera logic
+pub struct CameraPlugin;
+
+impl Plugin for CameraPlugin {
+    fn build(&self, app: &mut App) {
+        app;
+    }
+}