+++ template_lib/src/player/mod.rs
@@ -1,0 +1,3 @@
+//! Code that connects player character logic
+
+pub mod camera;