+++ RELEASES.md
@@ -3,6 +3,8 @@
 ## Version 0.1
 
 - Update template to `bevy` 0.10
+- Improve game architecture
+- added game basics for easy use of template
 
 ## Version 0.0